Convert IO to 64-bit all the time
To avoid confusion, use clear 64-bit types in the file I/O api.
Normally, 32-bit OS will define size and offsets as 32-bit instead of
64. Expect that we will be reading large files, and so make the API be
64-bit clean. This means that on a 32-bit os, the IO routines should use
things like lseek64 and similar, or return an error for a request that
is too large.

+++ src/lib/OpenEXRCore/openexr_read.h
@@ -183,7 +183,7 @@ typedef struct exr_decode_chunk_info__t
  * used to pre-allocate buffers for multiple threads in one block or
  * whatever your application may require.
  */
-EXR_EXPORT size_t exr_get_chunk_unpacked_size( exr_file_t *file, int part_index );
+EXR_EXPORT uint64_t exr_get_chunk_unpacked_size( exr_file_t *file, int part_index );
 
 /** @brief Computes the chunk_info for a particular scanline.
  *
